+++ artiq/firmware/libksupport/lib.rs
@@ -242,20 +242,20 @@ extern fn cache_put(key: *const u8, list: ArtiqList<i32>) {
 }
 
 extern fn i2c_start(busno: i32) {
-    send(&I2CStartRequest { busno: busno as u32 });
+    send(&I2CStartRequest { busno: busno as u8 });
 }
 
 extern fn i2c_stop(busno: i32) {
-    send(&I2CStopRequest { busno: busno as u32 });
+    send(&I2CStopRequest { busno: busno as u8 });
 }
 
 extern fn i2c_write(busno: i32, data: i8) -> bool {
-    send(&I2CWriteRequest { busno: busno as u32, data: data as u8 });
+    send(&I2CWriteRequest { busno: busno as u8, data: data as u8 });
     recv!(&I2CWriteReply { ack } => ack)
 }
 
 extern fn i2c_read(busno: i32, ack: bool) -> i8 {
-    send(&I2CReadRequest { busno: busno as u32, ack: ack });
+    send(&I2CReadRequest { busno: busno as u8, ack: ack });
     recv!(&I2CReadReply { data } => data) as i8
 }
